Changes for demo
To pick when the rover turns edit 'austin.xml'

If you want to play around with some of the model attributes, you can
edit them in 'config.xml'. Just a warning that a minor change of these
can have very large effects on the models but you wont break anything!

Both of these files are in /src/setup/